Change implementation from screen to popup
